[wavebar] Improve scale and calculation of sample position

+++ src/plugins/wavebar/settings/wavebarsettings.h
@@ -31,13 +31,13 @@ enum WaveBarSettings : uint32_t
 {
     Downmix       = 1 | Type::Int,
     ShowCursor    = 2 | Type::Bool,
-    CursorWidth   = 3 | Type::Double,
+    CursorWidth   = 3 | Type::Int,
     ColourOptions = 4 | Type::Variant,
     Mode          = 5 | Type::Int,
     BarWidth      = 6 | Type::Int,
     BarGap        = 7 | Type::Int,
     MaxScale      = 8 | Type::Double,
-    CentreGap     = 9 | Type::Double,
+    CentreGap     = 9 | Type::Int,
     ChannelScale  = 10 | Type::Double,
 };
 Q_ENUM_NS(WaveBarSettings)
